Importable Knowledge Connector App package
