+++ src/aggregator/index.ts
@@ -1,0 +1,1 @@
+export { DiffAggregator } from './diff-aggregator.js';
