+++ src/genesys/index.ts
@@ -1,0 +1,4 @@
+export { GenesysDestinationAdapter } from './genesys-destination-adapter.js';
+export { ExportArticlesResponse } from './model/export-articles-response.js';
+export { GenesysDestinationConfig } from './model/genesys-destination-config.js';
+export { configurer } from './configurer.js';
