+++ src/index.ts
@@ -1,19 +1,12 @@
-import { Pipe } from './pipe/pipe.js';
-import { GenesysDestinationAdapter } from './genesys/genesys-destination-adapter.js';
-import logger from './utils/logger.js';
-import { loadConfigurer } from './utils/configurer-loader.js';
-import { parseConfig } from './utils/config-parser.js';
+export { Config } from './config.js';
+export { Pipe } from './pipe/pipe.js';
+export { Configurer } from './pipe/configurer.js';
+export { Adapter } from './adapter/adapter.js';
+export { loadConfigurer } from './utils/configurer-loader.js';
+export { parseConfig } from './utils/config-parser.js';
+export { ExportModel, SyncModel } from './model/sync-export-model.js';
+export { SyncDataResponse } from './model/sync-data-response.js';
 
-const config = parseConfig();
-
-loadConfigurer(config)
-  .then((configurer) => {
-    return new Pipe()
-      .destination(new GenesysDestinationAdapter())
-      .configurer(configurer)
-      .start(config);
-  })
-  .catch((error) => {
-    logger.error('Connector app aborted.', error);
-    process.exitCode = 1;
-  });
+export * from './processor/index.js';
+export * from './aggregator/index.js';
+export * from './uploader/index.js';
