+++ src/model/sync-export-model.ts
@@ -8,6 +8,8 @@ import { DocumentAlternative } from './document-alternative.js';
 
 export interface SyncModel {
   version: number;
+  sourceId?: string;
+  readonlyContent?: boolean;
   importAction: {
     knowledgeBase: {
       id: string;
