+++ src/processor/index.ts
@@ -1,0 +1,3 @@
+export { HtmlTransformer } from './html-transformer.js';
+export { ImageProcessor } from './image-processor.js';
+export { PrefixExternalId } from './prefix-external-id.js';
