+++ src/salesforce.ts
@@ -1,0 +1,1 @@
+export * from './salesforce/index.js';
