+++ src/salesforce/index.ts
@@ -1,0 +1,3 @@
+export { SalesforceLoader } from './salesforce-loader.js';
+export { SalesforceAdapter } from './salesforce-adapter.js';
+export { configurer } from './configurer.js';
