+++ src/uploader/index.ts
@@ -1,0 +1,1 @@
+export { DiffUploader } from './diff-uploader.js';
