+++ src/zendesk.ts
@@ -1,0 +1,1 @@
+export * from './zendesk/index.js';
