+++ src/zendesk/index.ts
@@ -1,0 +1,3 @@
+export { ZendeskLoader } from './zendesk-loader.js';
+export { ZendeskAdapter } from './zendesk-adapter.js';
+export { configurer } from './configurer.js';
